// src/app/shared/interfaces/product.interface.ts
/******************************************************************************
 * Title: product.interface.ts
 * Author: Jeff Shepherd, Clayton Stacy, Chris Bohnet, Verlee Washington
 * Modified by:
 * Date: 10/28/2020
 * Description: product interface
 *****************************************************************************/

 export interface Product {
   _id: string;
   price: number;
   title: string;
 }